+++ app/src/components/FilterCard/reducer.js
@@ -1,4 +1,3 @@
-const TOGGLE_FILTER = 'TOGGLE_FILTER';
 
 const initialState = {
   electricCharger: false,
@@ -6,14 +5,21 @@ const initialState = {
   favorite: false,
   time: false,
   handicapped: false,
+  distance: 0,
 };
 
 export default function (state = initialState, action) {
   switch (action.type) {
-    case TOGGLE_FILTER:
+    case 'TOGGLE_FILTER':
       return {
         ...state,
         [action.filter]: !state[action.filter],
+      };
+
+    case 'UPDATE_DISTANCE_FILTER':
+      return {
+        ...state,
+        distance: action.value
       };
     default:
       return state;